feat: make a subset of the library independent from alloc
