+++ crates/rapier2d-f64/Cargo.toml
@@ -19,14 +19,21 @@ maintenance = { status = "actively-developed" }
 [lints]
 clippy = { needless_lifetimes = "allow" }
 rust.unexpected_cfgs = { level = "warn", check-cfg = [
-    'cfg(feature, values("dim3", "f32", "std", "simd-is-enabled", "simd-stable", "simd-nightly"))',
+    'cfg(feature, values("dim3", "f32", "std", "alloc", "simd-is-enabled", "simd-stable", "simd-nightly"))',
+    'cfg(target_arch, values("spirv"))',
 ] }
 
 [features]
 default = ["dim2", "f64", "std"]
 dim2 = []
 f64 = []
+alloc = [
+    "nalgebra/alloc",
+    "parry2d-f64/alloc",
+    "serde?/alloc",
+]
 std = [
+    "alloc",
     "parry2d-f64/std",
     "nalgebra/std",
     "simba/std",
@@ -69,7 +76,6 @@ required-features = ["dim2", "f64"]
 doctest = false # All doctests are written assuming the 3D version.
 
 [dependencies]
-nalgebra.workspace = true
 glamx.workspace = true
 parry2d-f64.workspace = true
 simba.workspace = true
@@ -83,6 +89,12 @@ static_assertions.workspace = true
 web-time = { workspace = true, optional = true }
 rayon = { workspace = true, optional = true }
 serde = { workspace = true, optional = true }
+
+# Nalgebra is only included on non-spirv targets. On spirv, rapier exposes only the
+# subset of its API that doesn't require nalgebra.
+[target.'cfg(not(target_arch = "spirv"))'.dependencies]
+nalgebra = { workspace = true }
+glamx = { workspace = true, features = ["nalgebra"] }
 
 [dev-dependencies]
 bincode.workspace = true
