+++ crates/rapier3d/Cargo.toml
@@ -19,14 +19,21 @@ maintenance = { status = "actively-developed" }
 [lints]
 clippy = { needless_lifetimes = "allow" }
 rust.unexpected_cfgs = { level = "warn", check-cfg = [
-    'cfg(feature, values("dim2", "f64", "std"))',
+    'cfg(feature, values("dim2", "f64", "std", "alloc"))',
+    'cfg(target_arch, values("spirv"))',
 ] }
 
 [features]
 default = ["dim3", "f32", "std"]
 dim3 = []
 f32 = []
+alloc = [
+    "nalgebra/alloc",
+    "parry3d/alloc",
+    "serde?/alloc",
+]
 std = [
+    "alloc",
     "parry3d/std",
     "nalgebra/std",
     "simba/std",
@@ -72,7 +79,6 @@ path = "../../src/lib.rs"
 required-features = ["dim3", "f32"]
 
 [dependencies]
-nalgebra.workspace = true
 glamx.workspace = true
 parry3d.workspace = true
 simba.workspace = true
@@ -87,6 +93,12 @@ wide.workspace = true
 web-time = { workspace = true, optional = true }
 rayon = { workspace = true, optional = true }
 serde = { workspace = true, optional = true }
+
+# Nalgebra is only included on non-spirv targets. On spirv, rapier exposes only the
+# subset of its API that doesn't require nalgebra.
+[target.'cfg(not(target_arch = "spirv"))'.dependencies]
+nalgebra = { workspace = true }
+glamx = { workspace = true, features = ["nalgebra"] }
 
 [dev-dependencies]
 bincode.workspace = true
