+++ src/control/mod.rs
@@ -1,16 +1,20 @@
 //! Utilities for controlling the trajectories of objects in a non-physical way.
 
+#[cfg(feature = "alloc")]
 pub use self::character_controller::{
     CharacterAutostep, CharacterCollision, CharacterLength, EffectiveCharacterMovement,
     KinematicCharacterController,
 };
+#[cfg(feature = "alloc")]
 pub use self::pid_controller::{PdController, PdErrors, PidController};
 
-#[cfg(feature = "dim3")]
+#[cfg(all(feature = "dim3", feature = "alloc"))]
 pub use self::ray_cast_vehicle_controller::{DynamicRayCastVehicleController, Wheel, WheelTuning};
 
+#[cfg(feature = "alloc")]
 mod character_controller;
 
+#[cfg(feature = "alloc")]
 mod pid_controller;
-#[cfg(feature = "dim3")]
+#[cfg(all(feature = "dim3", feature = "alloc"))]
 mod ray_cast_vehicle_controller;
