+++ src/data/coarena.rs
@@ -44,6 +44,7 @@ impl<T> Coarena<T> {
         self.data.get_mut(index as usize).map(|(_, t)| t)
     }
 
+    #[allow(dead_code)]
     pub(crate) fn get_gen(&self, index: u32) -> Option<u32> {
         self.data
             .get(index as usize)
