+++ src/dynamics/joint/impulse_joint/impulse_joint_set.rs
@@ -6,34 +6,7 @@ use crate::geometry::{InteractionGraph, RigidBodyGraphIndex, TemporaryInteractio
 
 use crate::data::Coarena;
 use crate::data::arena::Arena;
-use crate::dynamics::{GenericJoint, IslandManager, RigidBodyHandle, RigidBodySet};
-
-/// The unique identifier of a joint added to the joint set.
-/// The unique identifier of a collider added to a collider set.
-#[derive(Copy, Clone, Debug, PartialEq, Eq, Hash)]
-#[cfg_attr(feature = "serde-serialize", derive(Serialize, Deserialize))]
-#[repr(transparent)]
-pub struct ImpulseJointHandle(pub crate::data::arena::Index);
-
-impl ImpulseJointHandle {
-    /// Converts this handle into its (index, generation) components.
-    pub fn into_raw_parts(self) -> (u32, u32) {
-        self.0.into_raw_parts()
-    }
-
-    /// Reconstructs an handle from its (index, generation) components.
-    pub fn from_raw_parts(id: u32, generation: u32) -> Self {
-        Self(crate::data::arena::Index::from_raw_parts(id, generation))
-    }
-
-    /// An always-invalid joint handle.
-    pub fn invalid() -> Self {
-        Self(crate::data::arena::Index::from_raw_parts(
-            crate::INVALID_U32,
-            crate::INVALID_U32,
-        ))
-    }
-}
+use crate::dynamics::{GenericJoint, ImpulseJointHandle, IslandManager, RigidBodyHandle, RigidBodySet};
 
 pub(crate) type JointIndex = usize;
 pub(crate) type JointGraphEdge = crate::data::graph::Edge<ImpulseJoint>;
