+++ src/dynamics/joint/impulse_joint/mod.rs
@@ -1,6 +1,11 @@
 pub use self::impulse_joint::ImpulseJoint;
-pub use self::impulse_joint_set::{ImpulseJointHandle, ImpulseJointSet};
+pub use self::impulse_joint_handle::ImpulseJointHandle;
+#[cfg(feature = "alloc")]
+pub use self::impulse_joint_set::ImpulseJointSet;
+#[cfg(feature = "alloc")]
 pub(crate) use self::impulse_joint_set::{JointGraphEdge, JointIndex};
 
 mod impulse_joint;
+mod impulse_joint_handle;
+#[cfg(feature = "alloc")]
 mod impulse_joint_set;
