+++ src/utils/component_mul.rs
@@ -1,7 +1,9 @@
 //! ComponentMul trait for element-wise vector operations.
 
 use crate::math::Vector;
+#[cfg(not(target_arch = "spirv"))]
 use crate::utils::SimdRealCopy;
+#[cfg(not(target_arch = "spirv"))]
 use na::{Vector2, Vector3};
 
 /// Extension trait for element-wise vector operations
@@ -18,6 +20,7 @@ impl ComponentMul for Vector {
 }
 
 // Nalgebra vector implementations for SIMD support
+#[cfg(not(target_arch = "spirv"))]
 impl<N: SimdRealCopy> ComponentMul for Vector2<N> {
     #[inline]
     fn component_mul(&self, other: &Self) -> Self {
@@ -25,6 +28,7 @@ impl<N: SimdRealCopy> ComponentMul for Vector2<N> {
     }
 }
 
+#[cfg(not(target_arch = "spirv"))]
 impl<N: SimdRealCopy> ComponentMul for Vector3<N> {
     #[inline]
     fn component_mul(&self, other: &Self) -> Self {
